+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ Changelog
 Added
 ~~~~~
 * Added ``tenant`` argument to ``make_azure_blueprint``
+* Added ``hd`` argument to ``make_google_blueprint``
 
 Fixed
 ~~~~~
@@ -14,7 +15,6 @@ Fixed
 * Only set ``auto_refresh_url`` in ``make_google_blueprint`` if a token of
   type ``offline`` is requested. See issues `#143`_, `#144`_ and `#161`_ for
   background.
-* Added ``hd`` argument to ``make_google_blueprint``
 
 `1.0.0`_ (2018-06-04)
 ------------------