+++ CHANGELOG.rst
@@ -8,6 +8,7 @@ Added
 ~~~~~
 * Added ``tenant`` argument to ``make_azure_blueprint``
 * Added ``hosted_domain`` argument to ``make_google_blueprint``
+* Added Zoho pre-set configuration
 
 Fixed
 ~~~~~